fix cycle in treesitter land
